requirements.txt
alabaster==0.7.16
alembic==1.13.1
amqp==5.2.0
Babel==2.14.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
config==0.5.1
cryptography==42.0.5
docutils==0.20.1
Flask==3.0.2
Flask-Autodoc==0.1.2
Flask-SQLAlchemy==3.1.1
idna==3.6
imagesize==1.4.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
prompt-toolkit==3.0.43
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
